Refs #99421 Added resident role osse eligibility
