+++ app/models/resident_role.rb
@@ -7,6 +7,7 @@ class ResidentRole
   include Mongoid::Attributes::Dynamic
   include Mongoid::History::Trackable
   include GlobalID::Identification
+  include ChildcareSubsidyConcern
 
   RESIDENCY_VERIFICATION_REQUEST_EVENT_NAME = "local.enroll.residency.verification_request"
